Split Events.cs: Move objects into files, Move business logic into EventService
