+++ src/AasxServerAspNetCore/Program.cs
@@ -23,9 +23,8 @@ using Contracts;
 
 Console.WriteLine("AASX Server Core starting....");
 var host = CreateHostBuilder(args).Build();
-IPersistenceService persistenceService = (IPersistenceService)host.Services.GetService(typeof(IPersistenceService));
 
-AasxServer.Program.Main(args, persistenceService);
+AasxServer.Program.Main(args, host.Services);
 SecurityHelper.SecurityInit();
 await host.RunAsync();
 await host.WaitForShutdownAsync();
